Update README.md
lots of little changes and a few big ones too. I changed the wording of a couple system summaries and reviewed the "repairs and damage" section. Added a bit for the GBA SP and reworded some of the common problems bits. I copied the last section ("Modders") from the r/GameBoy wiki. I also merged the changes made to my wiki (r/game_boy) that differentiate it from r/gameboy